Stricter object type checking (#1920)
I realized that PODContent.fromEntries() wasn't checking its input type,
and that this can create weird results if someone passes the wrong type,
all because JavaScript is weird.

- Object.keys() returns an empty array when called on a number,
resulting in an empty POD
- Object.keys() returns ['0', '1', '2'...] (stringified digits, not
numbers) for a string or array, which still throws, but only because
digits are not valid PODNames.
- Furthermore I was reminded that arrays are a type of "object", so type
checking needs to be stricter to distinguish the two.

This PR enhances PODContent.fromEntries() to check its input type, and
enhances requireType() to more specifically distinguish arrays from
objects. I added new unit tests to catch the cases which wouldn't have
been caught by the old checks.

---------

Co-authored-by: Ahmad Afuni <root@ahmadafuni.com>