PODValue JSON Encoding (#1916)
Draft PR for feedback on the encoding format before I start using it
more broadly. The PR so far implements format and conversions for POD
values and entries of the 4 supported types. I want feedback on whether
this feels directionally correct, or if the unusual JSON format and
multiple encodings is something we might regret later. The test suite
isn't complete yet either, but I've tested enough to think the
implementation is complete. I didn't end up using any validator (Zod or
Valibot) because the existing checkPODValue does all the heavy lifting
already.

There are up to 3 potential JSON encodings for a given value, optimized
to allow for terseness without losing any type information (like the old
"simplified JSON" format did). When parsing, we always accept any of the
encodings, but when producing JSON we pick the simplest encoding which
fits the data without any loss.

Only ints and strings can use a type-free bare encoding. So e.g. an int
value 123 can just be the number `123`, while a larger int value needs
to be encoded as `{ "int": "12345678901234567890" }`.

We also parse the fully-explicit format `{ "type": "int", "value":
"12345678901234567890" }` but never produce it. It's redundant vs. the
terser version. I included it as a fallback, and because it's most
similar to the TypeScript PODValue, and could be simpler for some
clients to produce. I think it may be a candidate for removal.

What I didn't include is the in-string prefix encoding used in the old
"simplified JSON", like
`"eddsa_pubkey:c433f7a696b7aa3a5224efb3993baf0ccd9e92eecee0c29a3f6c8208a9e81d9e"`.
I'm on the fence about that. Removing breaks compatibility with existing
use cases. It's also slightly terser. On the flip side, though, it means
some valid PODStringValues can't be represented without a `pod_string:`
prefix as an escape-sequence, which seems like a footgun it might be
wise to do away with. E.g. when everything is a POD, what if someone
sets their password to `int:12345` and it gets misparsed as a number and
they can't log in?