Allow console.log statements to be dropped from builds if env var is set (#1950)
Closes
https://linear.app/0xparc-pcd/issue/0XP-1340/disable-logging-by-default-for-production-builds

Zupass logs a _lot_ of stuff, and those logs also appear in Zapps which
include Zupass via an iframe, which is annoying and confusing to third
party developers. We can disable logging in production by stripping
`console.log` statements from the build using `esbuild`.

This PR modifies the build script to strip `console.log` statements iff
an environment variable, `DISABLE_CONSOLE_LOG`, is set to `"true"`. If
we set this in production then logging would be disabled there.