Remove old pod serialization (#1952)
Removes the old serialization formats (json-bigint and simplified-json)
from POD library, and resolves all the necessary issues within the
Zupass repo. This is exploratory for now, we should discuss whether to
pull the trigger based on the details below.

2 things in Zupass needed non-trivial changes here:
- PODBox uses serialization to generate stable IDs in PODPipeline. This
PR does the minimal thing of updating that to use the new JSON format.
I'd like @robknight to confirm this.
- Minting uses old-form serialization without a PODPCD. This PR assumes
it will change to using a PODPCD wrapper. I think the pod-issuer server
is already capable of doing that if it uses the /api/sign URL instead of
the /api/mintPOD URL, but the URL is specified outside of Zupass. I've
submitted a draft PR in pod-issuer which I think will update this if we
decide it's necessary. I failed to get Deno to work in order to build
and test it, though. I'd like @ax0 to confirm this.

This is of course a breaking change for any external using these
serialization formats, which will need to be updated when there's a new
NPM release and they pick it up. I know of 3 such cases:
- ZuKyc uses POD serialization extensively, but the use cases are
straightforward. They'll need an update to use the JSON format, ideally
after there's a JSON format for GPC config as well.
- The LLM game (pod-arg) uses POD.serialize for signed API request. It's
a 1-line change it to use POD.toJSON
- The Minting server (pod-issuer) uses POD.serialize for minting as
described above. It also uses simplified-JSON serialization for its own
APIs, such as adding a POD. I've submitted a draft PR to change those to
the new format, though any existing input data would need to be
reformatted. I think Arturo is the only real user here, so we can work
with him if/when it becomes necessary.

I think there aren't any 3rd-party projects to worry about here, but am
open to reminders. The alternative here would be to make the Zupass
changes, but not remove the old serialization methods, simply marking
them as deprecated. I'm open to that, though I'm leaning toward just
ripping off the band-aid, so we don't have multiple confusing options
for developers to consider.