Expose functions to allow external provers/verifiers (#1986)
Splits out parts of gpcProve and gpcVerify into pre/post functions which
return circuit inputs/outputs. This allows the proving/verifying part to
be done separately, and thus replaced with another (compatible) system.
The specific use case for this immediately is on-chain verification, for
which gpcPreVerify should provide the necessary inputs. gpcPreProve and
gpcPostProve are unneeded for now, but could allow someone to plug in
and test a different proof stack if desried.

I also exposed the 3 key compiler functions behind these functions, in
case the pre/post functions don't fully meet the need. There's a TODO to
re-hide those if they aren't necessary in future.