Use patched Semaphore Proof package (#1987)
Closes
https://linear.app/0xparc-pcd/issue/0XP-1359/patched-version-of-semaphore-proof-315x

This PR introduces a patched version of `@semaphore-protocol/proof`
based on the `v3.15.2` tag of
https://github.com/semaphore-protocol/semaphore. This eliminates ~600KiB
of unused code from our bundle.

The only change is to how the `poseidon-lite` dependency is imported and
packaged:
1. `poseidon-lite` is listed as a dependency rather than
`devDependency`, which allows bundlers to bundle it correctly. The
previous misconfiguration meant that the build of
`@semaphore-protocol/proof` on NPM was over 600KiB in size, because it
included the entirety of the `poseidon-lite` package.
2. Import only the necessary sub-paths from `poseidon-lite`. This is
something that we have done in our imports from `poseidon-lite`, and is
also what `@semaphore-protocol/proof` v4 does. This just avoids
importing constants for Poseidon hash functions that are never used.

The code that the patched `@semaphore-protocol/proof` is published from
is here: https://github.com/robknight/semaphore/tree/dep-fix