Use JSON for GPCProofConfig and GPCRevealedClaims (#1983)
Make use of JSON serialization for GPCProofConfig and GPCRevealedClaims,
and delete the old serialization format. Tested with Consumer Client as
well as via unit tests.

Note that there is no backward-compatibility code in GPCPCD to parse the
old format, because we've never intended or promised backward
compatibility for GPC proofs (unlike PODs). The UnknownPCD fallback
handles any GPC proofs saved in Zupass (which is rare, but does happen
for people who've experimented with Consumer Client, like me and
Justin).

Depends on #1963.