JSON format for GPCs using Valibot (#1963)
Use Valibot to perform conversion to/from JSON for all GPC types
involved in serialization: GPCProofConfig, GPCBoundConfig, and
GPCRevealed claims, plus their sub-components. I did't cover
GPCProofInputs, which is more of a catch-all of multiple args, some of
which generally shouldn't be serialized (e.g. user identity), but all of
which do have some form of JSON serialization available individually.

This PR is limited to introducing the format, though I'll follow up with
a separate PR to make use of the format in GPC-PCD.

This PR also stops short at the amount of Valibot support needed for
these JSON conversions. I think there's an opportunity to let it take
over the "check" and "canonicalize" parts of the GPC compiler with some
more work, but that's a goal for the future.


Orignal comment for history of the start of conversation:

Results of experimentation with using Valibot to perform to-JSON and
from-JSON conversion for GPC configuration types. If this works out, it
might eventually replace the need for manual code in gpcChecks too. I
want feedback on the general pattern before I do the grunt work of
applying it to multiple types.

I think this is a solid approach, but I'm unhappy with the amount of
boilerplate involved, which could be missed when updating types in
future. Note that we don't need to declare or export parsers for all of
the intermediate types, maybe only for GPCConfig itself, so the
explosion of boilerplate need not be quite as bad as it seems.

There are two different representations of the same pattern to review
here. The code for BoundsCheckConfig in gpcValibot.ts is my initial hand
implementation. I think it uses the type system to catch as many
potential mistakes as possible. The code split out in
valibot/closedInterval.ts is an attempt to simplify the boiler plate by
putting it in a separate module, meaning no need for unique names. I
think it minimizes the amount of boilerplate which is actually variable
between different types, at the cost of more files, and a different
import syntax.

There are TODOs in the code marking my thoughts on the various parts of
this.

---------

Co-authored-by: Ahmad Afuni <root@ahmadafuni.com>