Fix missing client parameter for expiry of old cache entries (#2030)
When switching to using transactions, we started passing client objects
into functions that make SQL queries. `tryExpireOldEntries` is called on
a timer, without a parameter, so it doesn't receive a client object, and
it fails when calling lower-level DB access functions.

I've switched transaction opening around a bit, so that the transaction
is opened inside the function rather than outside of it, and changed the
only other place where this function is called to account for that.