Fix quirks of GPC artifact config and build (#2042)
Several things motivated by tracking down confusing behavior in my local
consumer-client build:

- When artifacts moved into a subdir by version, the build script wasn't
deleting the old dir, which lead to misconfigured code being able to
fetch the old files from the wrong URL.
- Env variable parsing for GPC config was built to assume "" and
undefined were equivalent, but not testing that, and not properly
enforcing it in some places.
- ZAPI-related prove/verify code wasn't respecting env variable config
because it bypassed the GPCPCD package.

I need @robknight to help me figure out how to test the ZAPI part of
this to make sure I didn't break it.