Return an error if the user cancels the connection (#2040)
This PR makes changes to how Zupass handles the Z API connection flow,
specifically so that if the user cancels by either cancelling the
opening of the initial popup, or by refusing permissions, an appropriate
message is sent to the connected app so that it can handle that error
state correctly.

In addition, a "Cancel" button has been added to allow the user to opt
out of connecting to Zupass at all, in line with the visual mocks.