podbox performance + caching (#2039)
closes
https://linear.app/0xparc-pcd/issue/0XP-1491/podbox-devcon-pipeline-risk-mitigation
#### caching
- new feature - persist cache of pipeline loads to disk
- this feature is enabled by the `LOCAL_FILE_SERVICE_ENABLED`
environment variable, and to enable it in prod we're going to need to
attach persistent volumes to our boxes
- pipelines by default persist their last successful load to disk, and
by default use that persisted load as a fallback when the pipeline is
paused or in an error state, unless caching is turned off for the
pipeline
- if a pipeline hasn't completed loading, and it has a cached load on
disk, it'll serve that
- if a pipeline experiences an error, it'll serve the last successful
load, even if it occurred 1+ restart(s) ago
- if a pipeline is paused, and has caching enabled, it'll serve the
cache
- it is possible to both clear and disable the persistent cache
- disabling the cache does not clear it, so that you can disable it, and
then re-enable it, and get back your old cache
- paused pipelines serve add + replace actions iff caching is enabled
and there's a cached load
- there should basically be no reason to ever disable caching or clear a
pipeline's cache, but i've added those features as a precaution
<img width="1470" alt="Screenshot 2024-10-25 at 3 10 34 AM"
src="https://github.com/user-attachments/assets/f206e349-213b-4531-91b4-1f151e863060">
<img width="1168" alt="Screenshot 2024-10-25 at 3 11 10 AM"
src="https://github.com/user-attachments/assets/be878178-b95d-4ac8-8568-70398ae98a91">
<img width="1181" alt="Screenshot 2024-10-25 at 3 11 49 AM"
src="https://github.com/user-attachments/assets/fb8d98b0-bd5e-47b1-9de3-0b77c55332df">
<img width="460" alt="Screenshot 2024-10-25 at 3 14 04 AM"
src="https://github.com/user-attachments/assets/5887ecef-4b9b-4310-a990-e0f5ec5021bb">

#### error handling
- product types defined in the pipeline definition that are not present
in pretix now cause a warning rather than an error, which can be used to
alert in discord. enabling this will require some tuning of the
'ignoredWarningsRegex' option in the Devcon pipeline definition
- a pipeline that has caching enabled, and has had a successful load
cached, is robust to errors that occur during any future loads

#### performance
- the Devcon pipeline details page now loads instantly in the Podbox
client - the problem was egregiously superfluous tracing code that
choked on the quantity of logs emitted by the Devcon pipeline load
- when querying pipeline atoms by email address, we use a hash map
rather than a list, which should speed things up
- the new `SELF_HOSTED_PODBOX_MODE` environment variable turns off
superfluous services (like `UserService` and `SemaphoreService` for the
self-hosted podbox situation

#### correctness
- the pipeline scheduler now correctly handles pipeline edits and
deletions. specifically, editing or deleting a pipeline now causes the
previous pipeline load to be aborted transparently to the user. in the
case of an edit, if the pipeline is not paused, it will be restarted,
again, transparently to the user. this means for e.g. if you pause the
devcon pipeline, it will not continue loading in the background (which
can take up to 4 minutes), and will wait for you to start it again. this
abort logic interacts properly with the scheduling logic - there will be
no loose pipelines running around.