Use wrapper package to import Semaphore v3 (#2019)
A couple of weeks ago, I changed how we import Semaphore v4, so that
rather than importing `@semaphore-protocol/core`, we directly imported
`@semaphore-protocol/identity@^4.0.3` under an alias. This avoided
loading some extraneous code from `@semaphore-protocol/core`'s unrelated
dependencies, and works fine for our use-case.

However, I discovered a scenario where this does not work well.

If some third-party application imports `@pcd/gpc`, then `@pcd/gpc` will
import some version of `@semaphore-protocol/identity`, according to what
is specified in its `package.json` file. If this third-party app manages
a Semaphore v4 identity, and wants to make a GPC proof using that
identity, then that app will _also_ import
`@semaphore-protocol/identity`, but may be importing a different
version.

Normally this is OK, if both dependencies are specified with the `^`
prefix - the bundler will know to select the highest version requested,
and satisfy both dependencies using that version. However, the aliased
import of `@semaphore-protocol/identity` breaks this, and so you can end
up with mismatched versions of `@semaphore-protocol/identity` being
installed.

This breaks GPC proof creation, because `checkProofInputs` does an
`instanceof` comparison between the `Identity` object provided as an
input, and the `Identity` class that it knows about. Since the input was
produced using one version of `@semaphore-protocol/identity`, but the
version that `@pcd/gpc` knows about is a different version, this check
fails. So far as I can tell, it fails even if there's no meaningful
difference between the class implementations. Strictly speaking, the
input is not an instance of the class it's expected to be.

I discovered this because for Z API testing purposes, I have a separate
"PARCNET client" which implements a subset of Zupass functionality,
including managing identities and making proofs, and this app suffered
from the issue described above. I think it's unlikely that any current
consumers of `@pcd/gpc` will be passing in Semaphore v4 identities to
their proofs, but this is functionality that certainly ought to work
out-of-the-box, and currently doesn't. It only works in Zupass because
Zupass happens to be using the same version of Semaphore v4 that
`@pcd/gpc` does.

The fix here is to avoid renaming the package. Instead, we can use a
wrapper package which simply imports everything from
`@semaphore-protocol/identity` and re-exports it. This gets us around
the problem of not being able to refer to two different versions of
`@semaphore-protocol/identity` in the same `package.json` file - we
simply refer to one of them indirectly. The dependency graph will
include two versions of `@semaphore-protocol/identity`, both with their
correct names. As the old joke goes, you can solve any problem in
software engineering by adding another layer of indirection - except for
having too many layers of indirection.

I've added a package which wraps `@semaphore-protocol/identity@3.15.2`
and re-exports it, and I've renamed the uses of `Identity` in `@pcd/gpc`
to `IdentityV3` and `IdentityV4` so that we're maximially explicit about
which types are coming from which package.