Fix insert/delete tracking (#2057)
When inserting or deleting PODs via the Z API, we track the change in
state of the storage revision number. However, we currently set up the
event listener immediately after making the change, which means there's
a race condition, in which the event might fire before the listener is
set up. This change reverses the order of operations, setting up the
event listener immediately before performing the insertion or deletion.