0XP-1251: horizontally scale zupass backend on single machine across multiple CPUs by running a nodejs cluster (#1981)
enabled by https://github.com/proofcarryingdata/zupass/pull/1948

#### summary

This PR horizontally scales the Zupass backend on a single machine
across multiple cores and processes. It does this via Node's
[cluster](https://nodejs.org/api/cluster.html) API, which enables a Node
application to fork itself and load balance incoming HTTP requests
amongst the child processes. I have *not* load balanced everything,
though I did load balance a significant portion of the app.

Here's how the clustering is implemented in this PR:
- on startup, the server checks if the 'clustering' feature is enabled
via the `ENABLE_CLUSTER` environment variable. if it's not, the server
functions precisely as it did prior to this PR.
- if clustering is enabled, the server forks (clones its own process)
itself once per CPU core, or a quantity specified by the
`CLUSTER_PROCESSES` environment variable. as a result there is one
'main' process, and some quantity of child processes.
- each of these processes (main and child) start their own instance of
the Zupass backend via the `startApplication` function. The main process
starts in a new 'mode' - `PARALLEL_MAIN`, and the child process Zupass
backends start in another new 'mode' - `PARALLEL_CHILD`.
- the `PARALLEL_CHILD` server instances listen to HTTP requests on a
different port from the main process (default `30002` vs. the main's
`3002`).
- the `PARALLEL_CHILD` server instances start only a subset of the
'services' that the Zupass backend hosts (examples of services include
`UserService`, `SemaphoreService`, etc.), whereas the `PARALLEL_MAIN`
process starts *all* of the services. The reason for this is that some
services are parallelizable and some are not (at least in their current
state).
- the `PARALLEL_MAIN` server (that runs on the main, initial process)
starts *all* of the services.
- to external clients, this parallelization is transparent. incoming
requests are always received by the main cluster process. depending on
the request, some of the requests are forwarded to one of the child
processes.
- the way i've implemented this forwarding is via a new Express.js
middleware -
[`clusterMiddleware`](https://github.com/proofcarryingdata/zupass/pull/1981/files#diff-b18d23998c8ba50c80b52f40f3656928a21e92475f0f298fc54a4078702c9e43).
I've added this middleware to *some* (*not all*) routes on the backend.
I've deliberately chosen to parallelize routes in this manner that would
not cause problems. For example, the `/account/salt` route is now
forwarded to a child process, rather than handled by a single 'main'
processes. [You can
see](https://github.com/proofcarryingdata/zupass/pull/1981/files#diff-5ef4b120db296036aae2210e294140d447e33a3748d16223c56932f04285cdf7R38-R52)
that all I had to do to make this happen on the route was to add the
`clusterProcess` middleware to the route's definition.

Here's the full list of relevant services/routes that are now
parallelized: `accountRoutes`, `e2eeRoutes`, `frogCryptoRoutes`,
`pcdIssuanceRoutes`.
- because `accountRoutes` are parallelized, actions like account
creation deletion and updates are parallelized.
- because `pcdIssuanceRoutes` is parallelized, e.g. the `EmailPCD` feed,
the one that all users must hit when creating an account and whenever
they load the Zupass client, is parallelized.

The notable services that are *not* parallelized in this PR are
`TelegramService` and `GenericIssuanceService`.
- `TelegramService` is not parallelized because the telegram API only
supports one connected client per bot. Parallelizing some of the
functionality is possible but probably beyond the scope of this PR.
- `GenericIssuanceService` is not parallelized for two main reasons: 1)
each pipeline's 'atoms' are stored in-memory, and Node's multiprocess
model does not facilitate efficient inter-process communication and 2)
each pipeline has a scheduled 'job' that loads 'atoms' from somewhere
(Pretix or Lemonade) on the internet. These are surmountable challenges,
but also beyond the scope of this PR imo.

Between the two notable non-parallelized services,
`GenericIssuanceService` is the more important. If necessary, I could
spec out what needs to be done to make that happen. I think now that
this architecture is in place, that task could be offloaded from me.