4 new PODValue types (#2046)
Adds 4 new PODValue types for convenience or special purposes
- bytes is a variable-size binary type isomorphic with strings
- boolean is a numeric type isomorphic to int with value 0 or 1
- date is a convenience type using JS Date, represented in-circuit as
int milliseconds, comparable to ints
- null is a special type which can be used to make entries explicitly
have no value, as an alternative to proving a value absent. It has a
hard-coded hash value to remain unique (assuming no hash collisions).

I decided against a uuid type since I wasn't convinced it had to be
numeric strongly enough to want to break other code, and as a
non-numeric type it has no value beyond string.

All these types are fully supported and checked in POD and GPC layers.
Tests at the GPC layer are somewhat bare-bones, but should benefit from
the fact that all the logic is centralized in a few functions in
@pcd/pod so if it works for one type it'll work for all. The biggest
assumption I had to break here was that "int" was the only "arithmetic"
type usable for range and inequality checks.

Closes:
https://linear.app/0xparc-pcd/issue/0XP-913/more-types-for-pod-values

---------

Co-authored-by: Ahmad Afuni <root@ahmadafuni.com>