Faster Z API connect by not waiting for all feeds to sync (#2120)
For discussion:

When connecting to an embedded Zupass, we currently wait for:
a) Downloading and extracting the e2ee storage
b) Fully processing all feeds

(b) can be slow as it involves additional network requests, and some
users may have several additional subscriptions beyond the "core" email
feed, Devcon feed, and so on.

Instead, we can wait for (a) as per usual, but if the e2ee storage
already contains a Devcon ticket, we can initiate the connection without
needing to wait for a full sync of the Devcon feed. This would allow
queries on the Devcon tickets to work, and make connection much faster
for all users.

The main risk is:
1) User logs in to Zupass, hits the feed, gets their ticket, saves it in
e2ee storage
2) User closes Zupass
3) User's Devcon ticket is deleted, amended, etc.
4) User signs in to a Zapp such as the PWA, gets the ticket from e2ee
storage
5) Zapp queries the ticket and sees the ticket loaded from e2ee storage
6) Within a few seconds, the user hits the Devcon feed and the ticket is
updated
7) Zapp might not notice that the ticket has changed

However, this case seems very unlikely to matter. To trigger it, you
would need to have a change made to your ticket, and not log in to
Zupass but instead log in to a Zapp. The issue would only occur on the
first time you log in to the Zapp, as the embedded Zupass would still
hit the feed and update the e2ee storage.

In addition, most Zapps don't have direct ticket access, so would be
unaffected.

Even for Zapps that _do_ have ticket access, it's not clear that the
effect would be particularly harmful. Some outdated data might be
visible on a single occasion, for the tiny (non-existent?) percentage of
users whose tickets change during the event.