Allow POD card display if at least one of title/description/image is set (#2152)
Logic changed since the original version of the POD card display so I've
been misinforming devs that they can make a collectable POD without an
image.

I tested and the UI code seems to work fine for text-only PODs, so this
PR loosens the check so that the collectable display can trigger if at
least one of the title/description/image entries are set to a non-empty
value.