Update @zk-kit/lean-imt (#2184)
Just a cleanliness update while I check for changes in our various
dependencies. Here's the story behind the updated comment, for
posterity:

I originally planned to use 0 as the hash value for PODNull, but decided
against it because LeanIMT behaved badly. Turns out that behavior was a
bug, fixed here:
https://github.com/privacy-scaling-explorations/zk-kit/pull/355

However, the test for that bug points out that people might assume that
0 has special meaning in many situations (including using it to mean
removal from a Merkle tree), so it's still a good idea to use a unique
non-zero value for PODNull's hash. In this PR I've updated the comment
on the related utest to describe the general reasoning, rather than
point to LeanIMT in particular.