Allow claiming of tickets without subscription (#2187)
Closes
https://linear.app/0xparc-pcd/issue/0XP-1697/improved-subscription-screen

This adds a new "Claim" screen, which is designed to work specifically
with Podbox feeds.

It polls the feed, extracting the POD Ticket PCD from the feed actions,
and then presents a preview of the ticket to the user. If the user
clicks "Claim", the ticket is added to their PCD collection.

This allows users to collect tickets without needing to subscribe to
feeds or to grant ongoing permissions for feeds to access folders,
including taking destructive actions like deleting PCDs. It also avoids
the overhead of recurring feed polling, both on the client and on
Podbox.

To test, get a feed link from Podbox in the same environment you're
testing in and then go to
https://zupass-url/#/claim?type=ticket&url=<feed url>

This should show a ticket preview, assuming that you have a ticket
available for your email address on the feed in question.