Add support for React 19 as a peer dependency (#2188)
[React 19 now has a stable
release](https://react.dev/blog/2024/12/05/react-19) and so we can
expect to see third-party apps using it.

Our packages specify that React ^18.2.0 is a supported peer dependency,
but this excludes React 19. In initial testing it seems that our
packages work with React 19, and so it should be listed as a valid peer
dependency.

The effect on third-party developers is that they will be able to use
our packages with React 19 without receiving warnings or errors from
their package manager, and without having to override those warnings or
errors (e.g. by using `--force` or `--legacy-peer-deps` with `npm`).

I've added a changeset for this, which can be used to version-bump the
affected packages when this PR is accepted.