Redirect if user not logged in when trying to claim ticket (#2192)
If the user is not logged in when trying to claim a ticket, redirect
them to the login/register screen, and redirect them back to the claim
after they sign in.

Also added the claim link to the Podbox UI, to make it easier for Podbox
admins to find claim links.