Use localForage instead of localStorage in most cases (#2195)
This PR switches out all uses of localStorage for localForage (backed by
IndexedDB) except for two:

1) In the one-click redirect flow, local storage is used to store the
redirect key. This does not share any code with our other uses of local
storage, which are centralized in localstorage.ts. I've decided to leave
this one alone, as otherwise it would involve changing the React
component flow to handle async access to IndexedDB.

2) A similar case occurs in Countdown.tsx in FrogScreens. This code
isn't used any more and can be removed.

The effect of the PR is that your existing login state, held in local
storage, will be ignored. Users with local storage state will be treated
as logged out, and must log in again. This will populate the IndexedDB
state with a fresh sync.