Compress the serialized PCD collection during sync if above a threshold size (#2197)
This PR amends the `serializeStorage` and `deserializeStorage` to apply
gzip compression to the serialized PCD string. By pre-compressing the
string, we reduce the size of the plaintext sync blob. This is helpful
as compression cannot be applied after the blob has been encrypted.

Testing with pseudo-random data indicates a reduction of 50% in the size
of the PCD collection string. We might hope to see better real-world
performance as real data will have more repetition (of names, email
addresses, public keys, and so on).