Force packages using Buffer to declare a dependency on it (#2199)
A few of our packages (but, most impactfully, `@pcd/pod`) make use of
the `Buffer` class without declaring a dependency on it. This does not
trigger build errors, because in NodeJS `Buffer` is a built-in global.
However, in browser environments it does not exist. This causes code
using `Buffer` to crash in browser environments.

A workaround for this is to install a polyfill. However, the polyfill
for Buffer is just the NPM `buffer` package, which we can declare a
direct dependency on. This improves the experience for the user of our
packages, who is not forced to work out how to install polyfills, and so
on.

I've added an eslint rule to prevent use of the global `Buffer` object,
so you now have to import it directly from a package you've declared as
a dependency. This flagged up multiple other examples of use of `Buffer`
without a declared dependency, so I've fixed those too.