chore(pcd/webauthn-pcd): fix linting error
The ESLint rule is warning us that we’re using the Node “Buffer” as a global. The simplest and most idiomatic fix is to import Buffer explicitly from the Node/buffer module, rather than relying on the global.

This way, we eliminate the usage of the global Buffer object and satisfy the no-restricted-globals ESLint rule.