Fix package dependencies (#2204)
The active fix here is the missing dependency of @pcd/gpcircuits on
@pcd/util. It was listed as a devDependency but used in production code.
At least in some packaging scenario (seen at a hackathon) this lead to
the BABY_JUBJUB_PRIME constant being undefined, leading to a type error.
I suspect in most packaging scenarios it is hidden because that same
dependency is pulled in by other packages.

I reviewed all the questionable dependencies of @pcd/pod,
@pcd/gpcircuits, and @pcd/gpc to see if they matched what was imported
in the src directory. That lead to another missing dependencies being
added, and to a few unnecessary dependencies being removed.

With this change I'm also starting to experiment with use of changesets
for version management. These changes are all marked as "patch".

Note that the tsconfig changes in other packages were auto-generated by
`yarn fix-references`.