Retroactive changeset for PR #2199 (#2207)
Retroactive changeset for PR #2199 

Based on a review of diffs since our last NPM package release, this is
the only change which affects POD/GPC packages, and deserves to be
mentioned in release notes. I'm trying to retroactively follow the right
process so that `yarn changeset` can be used to do a release manual
manipulation of version numbers.