NPM Package release (#2208)
The interesting part to review here is
[scripts/npm-package-release.sh](https://github.com/proofcarryingdata/zupass/pull/2208/files#diff-9a9932ffecfd35b0a8c856faff63c2718a307376cf74e1513dee16e499cc179c)
which replaces the old `new-release-*` scripts. I did my best to make it
idempotent and foolproof.

The rest is all automatic changes performed by `yarn changeset`.

Note that I think ideally this change would be pushed without squashing
to preserve the commit hash, but I don't have permissions to bypass the
merge queue on main. I'm not sure it actually matters when I'm skipping
the git tagging part of the publish workflow. For reproducibility, we
certainly wouldn't want other PRs to land on main before this one, which
would confuse the question of what was in a given release.